cors/access-control-basic-denied.html

<html>
<body>
 <script src="/resources/testharness.js"></script>
 <script src="/resources/testharnessreport.js"></script>
 <script src="support.js"></script>

<pre id='console'></pre>
<div  id=log></div>


<script type="text/javascript">
function log(message)
{
    document.getElementById('console').appendChild(document.createTextNode(message + "\n"));
}

if (window.layoutTestController)
    layoutTestController.dumpAsText();


var accessControlBasicDenied = function() {
    var xhr = new XMLHttpRequest;

    try {
        xhr.open("GET", CROSSDOMAIN + "/resources/access-control-basic-denied.php", false);
    } catch(e) {
        log("FAIL: Exception thrown. Cross-domain access is not allowed in 'open'. [" + e.message + "].");
        assert_true(false);
        return;
    }

    try {
        xhr.send();
        assert_true(false);
    } catch(e) {
        assert_true(true);
        //log("PASS: Exception thrown. Cross-domain access was denied in 'send'. [" + e.message + "].");
        return;
    }

    log(xhr.responseText);
};

test(accessControlBasicDenied, "access-control-basic-denied")

if (window.layoutTestController)
    layoutTestController.notifyDone();  
</script>
</body>
</html>
